synapse-to-sharepoint-complete/synapse-to-sharepoint-complete/requirements.txt
# Azure dependencies
azure-functions==1.18.0
azure-identity==1.15.0
azure-synapse-artifacts==0.17.0

# Database connectivity
pyodbc==5.0.1
sqlalchemy==2.0.25

# Microsoft Graph API
msal==1.26.0
requests==2.31.0

# Configuration and utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1

# Logging
structlog==24.1.0
